Move workspace executors to their own modules
