+++ tsrc/workspace/__init__.py
@@ -1,8 +1,6 @@
 """ Implementation of the tsrc Workspace: a collection of git repositories
 """
 
-import stat
-import textwrap
 from typing import cast, Iterable, List, Tuple, Dict, Any, Optional, NewType  # noqa
 
 from path import Path
@@ -14,6 +12,10 @@ import tsrc.git
 import tsrc.manifest
 
 from .manifest_config import ManifestConfig
+from .cloner import Cloner
+from .copier import FileCopier
+from .syncer import Syncer
+from .remote_setter import RemoteSetter
 
 
 class LocalManifest:
@@ -188,215 +190,3 @@ class Workspace():
     def get_url(self, src: str) -> str:
         """ Return the url of the project in `src` """
         return self.local_manifest.get_url(src)
-
-
-class Cloner(tsrc.executor.Task[tsrc.Repo]):
-    def __init__(self, workspace_path: Path, *, shallow: bool = False) -> None:
-        self.workspace_path = workspace_path
-        self.shallow = shallow
-
-    def description(self) -> str:
-        return "Cloning missing repos"
-
-    def display_item(self, repo: tsrc.Repo) -> str:
-        return repo.src
-
-    def check_shallow_with_sha1(self, repo: tsrc.Repo) -> None:
-        if not repo.sha1:
-            return
-        if self.shallow:
-            message = textwrap.dedent(
-                "Cannot use --shallow with a fixed sha1 ({repo.sha1})\n"
-                "Consider using a tag instead"
-            )
-            message = message.format(repo=repo)
-            ui.fatal(message)
-
-    def clone_repo(self, repo: tsrc.Repo) -> None:
-        repo_path = self.workspace_path.joinpath(repo.src)
-        parent, name = repo_path.splitpath()
-        parent.makedirs_p()
-        clone_args = ["clone", repo.url]
-        ref = None
-        if repo.tag:
-            ref = repo.tag
-        elif repo.branch:
-            ref = repo.branch
-        if ref:
-            clone_args.extend(["--branch", ref])
-        if self.shallow:
-            clone_args.extend(["--depth", "1"])
-        clone_args.append(name)
-        try:
-            tsrc.git.run_git(parent, *clone_args)
-        except tsrc.Error:
-            raise tsrc.Error("Cloning failed")
-
-    def reset_repo(self, repo: tsrc.Repo) -> None:
-        repo_path = self.workspace_path.joinpath(repo.src)
-        ref = repo.sha1
-        if ref:
-            ui.info_2("Resetting", repo.src, "to", ref)
-            try:
-                tsrc.git.run_git(repo_path, "reset", "--hard", ref)
-            except tsrc.Error:
-                raise tsrc.Error("Resetting to", ref, "failed")
-
-    def process(self, repo: tsrc.Repo) -> None:
-        ui.info(repo.src)
-        self.check_shallow_with_sha1(repo)
-        self.clone_repo(repo)
-        self.reset_repo(repo)
-
-
-Copy = NewType('Copy', Tuple[str, str])
-
-
-class FileCopier(tsrc.executor.Task[Copy]):
-    def __init__(self, workspace_path: Path) -> None:
-        self.workspace_path = workspace_path
-
-    def description(self) -> str:
-        return "Copying files"
-
-    def display_item(self, item: Copy) -> str:
-        src, dest = item
-        return "%s -> %s" % (src, dest)
-
-    def process(self, item: Copy) -> None:
-        src, dest = item
-        ui.info(src, "->", dest)
-        try:
-            src_path = self.workspace_path.joinpath(src)
-            dest_path = self.workspace_path.joinpath(dest)
-            if dest_path.exists():
-                # Re-set the write permissions on the file:
-                dest_path.chmod(stat.S_IWRITE)
-            src_path.copy(dest_path)
-            # Make sure perms are read only for everyone
-            dest_path.chmod(0o10444)
-        except Exception as e:
-            raise tsrc.Error(str(e))
-
-
-class RemoteSetter(tsrc.executor.Task[tsrc.Repo]):
-    def __init__(self, workspace_path: Path) -> None:
-        self.workspace_path = workspace_path
-
-    def quiet(self) -> bool:
-        return True
-
-    def description(self) -> str:
-        return "Setting remote URLs"
-
-    def display_item(self, repo: tsrc.Repo) -> str:
-        return repo.src
-
-    def process(self, repo: tsrc.Repo) -> None:
-        try:
-            self.try_process_repo(repo)
-        except Exception as error:
-            raise tsrc.Error(repo.src, ":", "Failed to set remote url to %s" % repo.url, error)
-
-    def try_process_repo(self, repo: tsrc.Repo) -> None:
-        full_path = self.workspace_path.joinpath(repo.src)
-        rc, old_url = tsrc.git.run_git_captured(
-            full_path,
-            "remote", "get-url", "origin",
-            check=False,
-        )
-        if rc == 0:
-            self.process_repo_remote_exists(repo, old_url=old_url)
-        else:
-            self.process_repo_add_remote(repo)
-
-    def process_repo_remote_exists(self, repo: tsrc.Repo, *, old_url: str) -> None:
-        full_path = self.workspace_path.joinpath(repo.src)
-        if old_url != repo.url:
-            ui.info_2(repo.src, old_url, "->", repo.url)
-            tsrc.git.run_git(full_path, "remote", "set-url", "origin", repo.url)
-
-    def process_repo_add_remote(self, repo: tsrc.Repo) -> None:
-        full_path = self.workspace_path.joinpath(repo.src)
-        tsrc.git.run_git(full_path, "remote", "add", "origin", repo.url)
-
-
-class BadBranches(tsrc.Error):
-    pass
-
-
-class Syncer(tsrc.executor.Task[tsrc.Repo]):
-    def __init__(self, workspace_path: Path) -> None:
-        self.workspace_path = workspace_path
-        self.bad_branches = list()  # type: List[Tuple[str, str, str]]
-
-    def description(self) -> str:
-        return "Synchronize workspace"
-
-    def display_item(self, repo: tsrc.Repo) -> str:
-        return repo.src
-
-    def process(self, repo: tsrc.Repo) -> None:
-        ui.info(repo.src)
-        repo_path = self.workspace_path.joinpath(repo.src)
-        self.fetch(repo_path)
-        ref = None
-
-        if repo.tag:
-            ref = repo.tag
-        elif repo.sha1:
-            ref = repo.sha1
-
-        if ref:
-            self.sync_repo_to_ref(repo_path, ref)
-        else:
-            self.check_branch(repo, repo_path)
-            self.sync_repo_to_branch(repo_path)
-
-    def check_branch(self, repo: tsrc.Repo, repo_path: Path) -> None:
-        current_branch = None
-        try:
-            current_branch = tsrc.git.get_current_branch(repo_path)
-        except tsrc.Error:
-            raise tsrc.Error("Not on any branch")
-
-        # FIXME: is repo.branch allowed to be None ?
-        if current_branch and current_branch != repo.branch:
-            self.bad_branches.append((repo.src, current_branch, repo.branch))  # type: ignore
-
-    @staticmethod
-    def fetch(repo_path: Path) -> None:
-        try:
-            tsrc.git.run_git(repo_path, "fetch", "--tags", "--prune", "origin")
-        except tsrc.Error:
-            raise tsrc.Error("fetch failed")
-
-    @staticmethod
-    def sync_repo_to_ref(repo_path: Path, ref: str) -> None:
-        ui.info_2("Resetting to", ref)
-        status = tsrc.git.get_status(repo_path)
-        if status.dirty:
-            raise tsrc.Error("%s dirty, skipping")
-        try:
-            tsrc.git.run_git(repo_path, "reset", "--hard", ref)
-        except tsrc.Error:
-            raise tsrc.Error("updating ref failed")
-
-    @staticmethod
-    def sync_repo_to_branch(repo_path: Path) -> None:
-        try:
-            tsrc.git.run_git(repo_path, "merge", "--ff-only", "@{u}")
-        except tsrc.Error:
-            raise tsrc.Error("updating branch failed")
-
-    def display_bad_branches(self) -> None:
-        if not self.bad_branches:
-            return
-        ui.error("Some projects were not on the correct branch")
-        headers = ("project", "actual", "expected")
-        data = [
-            ((ui.bold, name), (ui.red, actual), (ui.green, expected)) for
-            (name, actual, expected) in self.bad_branches
-        ]
-        ui.info_table(data, headers=headers)
-        raise BadBranches()
